+++ example/example/views.py
@@ -1,6 +1,0 @@
-from django.shortcuts import render
-
-
-def page_view(request, route):
-    """Handle basic 'Page' requests."""
-    return render(request, 'example/page.html', {'route': route})